Add files via upload
These files are the extent of our work for MariHacks this year. We had originally aimed to create a much more comprehensive algorithm for calculating a more effective rating, but as seen in the code, it didn't turn out like I wanted it too. In the end, I came to the conclusion that while we could still attain a reasonable-ish product, the most effective version of our idea would have been developed in a language that has more comprehension in terms of loops and conditional statements compared to python. 

Thank you for checking out our project!